requirements.txt
# API
fastapi
uvicorn
pydantic
requests

# Gen AI 
openai
google-genai
google-cloud-aiplatform

# Video, Audio Processing
mutagen
moviepy
manim
gtts
TTS
opencv-python